Report & Results/requirements.txt
streamlit
PyPDF2
python-docx
pytesseract
Pillow
groq
python-dotenv